pyproject.toml
[tool.poetry]
name = "stubborn"
version = "0.1.0"
description = "Just a simple REST API stub service."
authors = ["Sergei Nagaev <nagaev.sv@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
django = "3.2.6"
djangorestframework = "3.12.4"
whitenoise = "5.3.0"
uwsgi = "2.0.19.1"
django-environ = "0.4.5"
psycopg2 = "2.9.1"
argon2-cffi = "20.1.0"
django-extensions = "^3.1.3"
django-nested-admin = "^3.3.3"

[tool.poetry.dev-dependencies]
mypy = "0.910"
django-stubs = "1.8.0"
coverage = "5.5"
django-coverage-plugin = "2.0.0"
flake8 = "3.9.2"
flake8-isort = "4.0.0"
pydocstyle = "6.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
